Fix streamstats getter

+++ pkg/acapapp/frameProvider.go
@@ -245,7 +245,7 @@ func (fp *FrameProvider) Stats() (*FrameProviderStats, error) {
 		Bitrate:                       m.GetUint32("bitrate", 0),
 		BufferType:                    m.GetString("buffer.type", ""),
 		Channel:                       m.GetUint32("channel", 0),
-		Format:                        m.GetUint32("format", 0),
+		Format:                        m.GetInt16("format", 0),
 		Framerate:                     m.GetUint32("framerate", 0),
 		GOPLength:                     m.GetUint32("gop_length", 0),
 		H26xIntraRefresh:              m.GetUint32("h26x.intra_refresh", 0),
@@ -277,7 +277,7 @@ func (fp *FrameProvider) Stats() (*FrameProviderStats, error) {
 		StatisticsLastFrameTS:         m.GetUint64("statistics.last_frame_ts", 0),
 		StatisticsReclaimCount:        m.GetUint32("statistics.reclaim_count", 0),
 		Width:                         m.GetUint32("width", 0),
-		ZipProfile:                    m.GetUint32("zip.profile", 0),
+		ZipProfile:                    m.GetInt16("zip.profile", 0),
 	}
 
 	return &FrameProviderStats{
